% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isoraster.R
\name{isoline_terra}
\alias{isoline_terra}
\alias{isoband_terra}
\title{Iso bands and iso lines for terra}
\usage{
isoline_terra(x, line)
}
\arguments{
\item{x}{a terra raster, only first layer is considered}

\item{line}{a contour level for isoline_terra}

\item{lo}{a polygon level interval minimum for the low values (must have a corresponding hi value)}

\item{hi}{a polygon level interval maximum for the high values (cf. lo)}
}
\value{
SpatVector object, for isoline_terra a table of spatial lines with 'line' contour value,
for isoband_terra a table of spatial polygons with 'lo' and 'hi' contour values
}
\description{
Produce contours for a raster and return a vector dataset.
}
\details{
\code{isoline_terra} returns lines, use 'line' to specify or an automatic set will be computed

\code{isoband_terra} returns polygons, use 'lo' AND 'hi' to specify breaks, or an automatic set will be computed

These lines and boundaries are nice because isoband uses marching squares, not literal pixel outlines.
}
\examples{
isoline_terra(rast(volcano))
isoband_terra(rast(volcano), lo = 100, hi = 150)
}
